Change Slash theme submodule repo to own forked version
It turns out that the Slash theme is missing a couple of things that I'd
like, so I'm going to fork it, add the things I want, and send pull
requests. Until my pull requests get merged, it looks like I'll be
maintaining my own forked version of the theme.

  https://github.com/lencioni/Octopress-Theme-Slash