Improve styles
While working on my first blog post, I noticed a few styles that I
wanted to tweak slightly. Most of the styles here make subtle
improvements to readability and design on top of the Slash theme. Some
future work could probably be done to try and get some of these styles
merged upstream, but in the meantime, this gets the job done.