Fix "read more" link
I had tried to use the special "more tag" to keep my homepage to a
reasonably browsable length, but had accidentally included an extra dash
when using the special comment. After it was working, I noticed that my
tweaks to the line-height needed to be applied to this element as well.