Add anchors for inline links to CSS post
I've been finding myself wanting to directly link to specific sections
in this article. Since Octopress does not add anchors or IDs to headings
automatically, I'm adding some anchors to this post as a quick fix to
allow for inline linking.

It may be worth investigating whether there is a plugin for Octopress
that adds this automatically, and if one does not exist, creating one or
contributing to Octopress directly.