Help Google understand who I am
I'd like for my author information to appear next to Google search
results. For this to happen, Google needs to have a pretty high degree
of confidence that my Google+ profile and my blog posts are intended to
be linked.

They outline two options to make this work at:

  https://support.google.com/webmasters/bin/answer.py?hl=en&answer=1408986

Since I haven't set up email on this domain, I opted for Option #2,
which is to add `?rel=author` to my Google+ link from my about page, and
to add a reciprocal link back from my Google+ page to my blog, under the
"Contributor To" section.