Update sass-enhance article for Bower/Rails integration
Since writing this article, we have improved our Bower/Rails integration
at Causes, and I wrote a blog post about it. This commit updates a
mention of our Bower/Rails integration method and adds a link to the
blog post that describes what we did.

At the same time, I noticed and fixed a typo, changing "ear" to "near".