Fix code in Sass color chips post
I recently upgraded to Sass 3.3 on a project where I was using this
code, and it ended up failing on a line where a variable was passed to a
function. Changing it so that the value is the result of
`node.expr.perform` instead of just `node.expr` fixed it. This is what
Sass does internally anyway, so it seems better.